Reworked the setupMenu() of datanav Form. The end result is basically the same, but it works better now for an app of mine that wants to add a menu after the Actions menu.
Expanded on some of the dMenuBar docstrings, and added a new function
getMenuIndex(), which returns the index of the menu with the passed caption,
or None. Also added a remove() function which removes the menu at the
specified index. Modified getMenu() to return None if the menu isn't found,
instead of the wx-default of returning -1.