Added DynamicEnabled property to dMenuItem. This property takes a function that should return True/False to determine if the item should be enabled/disabled when the parent menu is opened.
Added an example of this in the Actions menu item "Quick Report", 
which will be disabled before a query has been run, and enabled
afterward.

In working on the above, I needed to fix a big problem with dabo
menu items: when they are added to a dMenu using wx functions, the
actual menu item reverts back to a pure wx object with no remnants
of our dPemMixin anymore. I worked around this by maintaining a 
dictionary which maps the wx object to the dabo object.

TODO still: Make a dMenuSeparator object, because as it stands
any separators will be listed in dMenu.Children, but as pure wx
objects and not Dabo objects.

Also, we should get DynamicEnabled implemented in dToolBar, but
a quick look leads me to believe that toolbar items aren't 
objectified like dMenuItems are.