Added a new class named 'DataSet', which is a subclass of tuple. The purpose of this class is to allow SQL-like selections and joins to be made on data as it exists in Dabo, instead of strictly on the backend. For example, let's say you create two cursors, and then want to do a join on the data in them. In Fox it's no big deal, but you currently can't do that in Dabo without calling back to the database and re-fetching the data.
As of now there is a *very* rudimentary 'select()' method for this class. As time goes on, I'd like to flesh it out a bit more, and then eventually create a join() method to go along with it that would accept another DataSet object along with a bunch of conditions, and return the appropriate result set.