Preliminary changes to dToolBar. I'm going to be making a dToolBarItem class and fully dabo-objectifying the interface, and adding such functions to dToolBar as remove(), getItem(), getItemIndex(), insert(), prepend() etc. IOW, pretty much mirroring dMenu which makes sense because there are plenty of similarities between menus and toolbars.
The notable feature of this change is that you can now query dToolBar's
Children property. You get raw wx toolbar items, but that's better than
the empty list that wx returns by default.

The NextID functionality isn't needed because wx provides a safe function
for that already, so I removed it.