Implemented joins between two DataSets. Right now you can specify left, right and inner joins; I still have to work on outer joins. I also want to add the ability to alias a field name, and to automatically munge a duplicate field name (i.e., if both DataSets have a field named 'city', the field in the DataSet being joined will be changed to 'city_alias', where 'alias' is the alias specified for that DataSet.
This is going to make programming with Dabo so much more powerful, with much of the power that Fox brings: being able to get data from backend data stores, and also being able to manipulate local cursors without having to hit the backend. And by making the syntax as SQL-like as possible, it will be a no-brainer for developers to use.