OK, this should just about do it for the integration of SQLite into the DataSet class. It is working well enough to pass every test I've thrown at it. There may be additional data types that will need to be handled, but both datetime and Decimal are working properly.
First change: renamed the select() method to 'execute()', since it can now execute any SQL statement, not just selects. You can do updates, inserts and deletes, too. In those cases, the execute() operates on the current DataSet object's data, and the modified DataSet is returned.

Joins are now fully supported. To join several DataSets together, you write your SQL code using 'dataset' as the alias for the DataSet object on which you call execute(), and any other legal alias names for the other DataSets in the query. Then you create a dict with the alias name for a given DataSet as the key, and the DataSet itself as the value. You then pass that dict as the second parameter to execute(), and it will resolve the aliases appropriately.