Lots of tweaks and improvements to the SQLite engine in the DataSet class. I added a sort() method that was intended as a replacement for the current sort routine in the dCursorMixin class, but it turned out to be about 10 times slower. The problem is the constant loading/unloading of the data from a Python list of dicts to a SQLite table, and then converting back to list of dicts.
However, this got me thinking: what if the data wasn't constantly loaded and discarded? So I added code to persist the data in the SQLite table as long as the DataSet remains in existence. There is still the issue of changes from the app being propagated back to the DataSet, but I'm planning on looking into this next. There is also an opportunity, I think, to move the population of these DataSet tables into a separate thread, so that it can be done in the background without hanging up the app. 

Added adapter/converter methods for dMementos. First I tried pickling them for storage, but that was horribly slow, so I just store them locally and add them back to the result set. I also began working on code to change the name of the field in the SQLite table from 'dabo-' to 'dabo_', since hyphens are not legal in field names. This is halfway-done, as I also have to fix this in the dataset population routine.

Speaking of which, I made several speed improvements after spending some time on the pysqlite site. I switched to named parameters instead of positional ones, since that allows me to pass a dict instead of having to extract the dict.values() first. I also switched to using executemany() to do the insert by adding a generator function to generate the rows. Finally, I found the code to change the result from tuples to the type of DictCursor that we use in Dabo, saving more time re-constructing the dict for each row manually.