Added a 'Name' property to the Column class. This is a read-only prop that returns "col_%s" % self._dataField. It is intended only so that columns display properly in the Designer. When working with grids, you should continue to rely on the DataField as the identifier.
Added the '_addEmptyRows()' method that will add some dummy blank data to a grid so that it displays better. To use it, simply set grid.emptyRowsToAdd to a value > 0 before calling fillGrid().

Fixed a bug where an error would be raised if the user clicked on the header of a column when there was no data in the grid.