Corrected Paul's indentation correction to dFormMixin. Man, I committed the first copy while rushing out so that I could get it onto my notebook, because I wouldn't have net access for a few hours. First thing I do is run a test, and find the problem.
Added an option to the Sizer setter in dPemMixin to not destroy the current sizer that is set as the object's Sizer. By passing None to the property, the old sizer is no longer associated with the object, but it is not automatically destroyed.

Added the HGap and VGap properties to the dGridSizer class.