Implemented the tracking of an object's controlling Sizer and SizerItem rather than relying on wxPython's built-in discovery, since it didn't work for sizers inside of other sizers, and this was critical for proper restoration of the objects saved in the Designer when creating the form from the cdxml file.
The form class has been enhanced to use the new information to be able to re-create more complex forms that include sizers embedded in other sizers. There is still some work to be done with properly saving/restoring some of the Grid Sizer properties, but in general things are working pretty well.