Added the method 'getConnectionNames()' to the app, which will return the set of all currently-defined connections (open or not).
Added the method 'setdFormClass()' to dabo.ui.uiwx.__init__.py. This accepts either 'MDI' or 'SDI' as a parameter, and will change the definition of 'dForm' to either the MDI or SDI version.

Updated the dSplitForm class to use the dForm defined in the module instead of the version defined in the dForm.py file. Cleaned up some ancient code and spacing in both dSplitForm and dSplitter, and updated their test cases.