Added constants needed to distinguish the Find/Replace dialog fields from each other.
Added the method 'addConnectFile()' to dApp. This allows you to pass a path to a cnxml file, and it will read in the connection information and add it to its dbConnectionDefs.

Modified the dToolForm class to use wx.MiniFrame instead of a modified wx.Frame. I don't notice any change in behavior, but I figured that since this wx class is designed for just this purpose, it made more sense to use it. I left the old dToolForm code commented in case we find any problems with this new class.

Overhauled the Find/Replace logic in uiApp. It now works as expected (at least on OS X!), and is ready to be used as an integral part of the next iteration of the Designer, which will feature code editing.