This is a test of the new MDI dynamic class binding. dForm's __init__ manually tweaks dForm.__bases__ depending on whether dabo.settings.MDI is currently set to True or not. I'm committing this now to see if it causes issues on Windows and Mac.
I had to tweak autoBindEvents to stop crawling up the parent hierarchy and 
go directly to self.Form because while in MDI mode, one of the parents is a 
raw ws.MDIParentWindow which doesn't have a Parent property.

Because of the above change, I had to give dColumn a Form property, and the
easiest way to do that was to have it descend from dPemMixinBase instead of
dObject.

Assuming this works on Mac and Windows, I'll finish the job by removing the 
commented code in dForm and doing similar changes to dFormMain.