Changed dabo.icons.getIconFileName to return None if the icon path doesn't exist. Before, it would return a fully-qualified bogus path if the filename didn't exist. Modified dIcons.getIconBitmap() to cope with this change.
Modified Icon property of dFormMixin to reliably take either a wx.Icon or 
a icon name, and convert if necessary to a wx.Icon. But I made the dabo
property keep the string name if that was what was passed, which I think
is better idea than referencing the wx.Icon reference. For instance, saving
a value of 'daboIcon.ico' from the Designer is much easier to deal with than
saving a value of '<wx.Icon instance at C0000006>' <g>.