Removed the yellow background from the dialog - that was just for debugging.
Removed the split sash code from the dEditor. That stuff belongs in the form
the editor is located, not in the editor itself.

Fixed a bug in dGrid that only seemed to pop up when the grid was located on
a form with a bizobj but that wasn't a dForm. The presets dialog in
minesweeper fit that description, and it was throwing errors when you
navigated through the grid.