Okay, since we have that pnlControl in the way, we need to add objects to the panel and then use the panel's sizer. And because I was relying on the objects being referenced from the form, I instead changed to using the RegID instead of Name for those objects.
I think it is working again, and I like the layout better - nicer use
of borders.

However, I'm still resisting the user having to remember about the 
existence of self.pnlControl. They should be able to add their objects
to self, and to use self.Sizer, since that is how anyone would try it
at a first shot. IOW, self.pnlControl may be a necessary implementation
detail, but the interface to the user should still be via self. If the
user wants to add their own panels, and subpanels of those, they should
be able to do that, but via self, not self.pnlControl.

Sorry if I'm being a PITA on this.