Modified the class structure for the calendar class. Now instead of having to specify the 'CompactCalendar' property to get the more common calendar, this type is now named 'dCalendar', while the version with the separate month and year controls is named 'dExtendedCalendar'.
Added a form subclass named 'dBorderlessForm'. This creates a form with no visible border or title bar. Renamed the base form class in dForm.py from 'dFormBase' to 'BaseForm'. I felt that the 'd' prefix was misleading, as it isn't meant to be a Dabo base class at all.

Started writing code for dabo.ui.popupCalendar(), but couldn't figure out how to simulate a modal popup without using a dialog instead of a frame-based control. It's commented out for now.