Added better sizing of the contents of the panels when the dFoldPanelBar is in Singleton mode. This allows for the contents of the expanded panel to fill the available space better.
Added a CaptionHeight property to wrap the wx call to the caption bar's GetSize() method. Added a CaptionForeColor property, but this doesn't work in the current release of the wx.lib.foldpanelbar.py script. I've submitted a patch to the wxPython list.

Oh, yeah - fixed a bug in my earlier dGrid bug fix. <sigh!>