dGrid: Began working on a problem when the grid has a DataSource specified that is not a bizobj. I got some of the problems fixed, but it seems that sorting doesn't work yet, and I will have to test further to find out why.
Removed the RegID from the incremental search timer added to each grid. If you have more than one grid on a form, it will cause an error.

Added the method 'getColByX()' to return a column reference instead of just a column number.

Added an attribute '_vetoAllEditing' to the grid that will disallow editing even if the grid and column allow it. This is because I don't want to be able to edit the grid in the Designer, but still want to mark the fields accordingly.


dFormMixin:
Added code to allow it to correctly re-create a grid at runtime from the cdxml file. Also added some code to fix a problem where the xmltodict function seems to be mangling square brackets in CDATA sections.