Changed the MenuOpen from a MouseEvent to a MenuEvent.
Added basic MRU (Most Recently Used) capabilities to Dabo. I had looked at the wx.FileHistory class, but it seemed a bit too rigid to work with our design. 

To use MRU in a menu, the menu should be created by passing 'MRU=True' in the constructor. I've changed the File menu in dBaseMenuBar to always be MRU capable, so since that's where 99% of the MRU stuff will be applicable, this should take care of most cases.

To add an item to MRU, call self.Application.addToMRU(), passing the menu in question, the menu prompt (typically the file path), and the function to be bound to that MRU item. If that particular prompt isn't yet in the MRU list, it will be prepended to the top. If it is already in the list, it will be moved up to the top.

Right now this only works with menubar-type menus; i.e., those that have a Caption. The MRU lists are kept distinct based on Caption, since multiple forms may share a menu, and a file opened from one form should appear in the MRU of all those forms. Since context menus don't have a Caption, they won't work yet, but I have an idea for passing an optional identifier to use in such cases.

Also updated the dMenu class to return the added separators, menuitems, etc. Most of the methods were not returning anything.