Added some tests to doDefaultMixin.
NOTE: I was *wrong*. We must call doDefault() on the class, not on self. It
is because self's __class__ always refers to the youngest subclass, and not
the proper superclass. So, while self.doDefault() does work if you never
make a subclass of the class where you use it that way, it can't be used
because as soon as you subclass that, it will go boom.