Fixed a quirk in how Windows handles menu events.
Added 'removePage()' and 'cyclePages()' functions to the dPageFrameMixin class. The former allow for a specific page to be removed, while the latter accepts an integer value and cycles the selected page forward (positive) or backwards (negative).

Added a clear() function to dMenu() for removing all current items.

Improved the default file name handling for dEditor.