Added the TreeItemContextMenu event. This differs from the normal ContextMenuEvent in that it returns the itemID of the node that was right-clicked.
Added code in the event data gathering code to store the tree item ID for these events in EventData.itemID.

Changed all the names '_id' in the dTreeView.py file to the clearer 'itemID'.

Fixed an oversight in the dSizerMixin that didn't remove the Box for a dBorderSizer when the sizer was released.