Major revamping of the whole refresh/update/dynamic design.
First, refresh() simply repaints the object and anything contained in it. It simply calls the wx.Refresh() method.

There is now a method named 'update()' which will update any dynamic aspects of a control, such as its Dynamic* properties, or its data bindings, if any. If the object is a form, calling update triggers a ValueUpdate event that all data-aware controls will receive. All objects with children will raise an Update event in order to propagate the update message. You can call update() on a specific object, and only that object and its children will be updated.

This clarifies the language we use: 'refresh' refers only to having the object repainted on the screen, while 'update' can change aspects of the object, such as its value or other properties.

Updated all the classes that referred to these events and methods by their old names.