Simplified dPemMixin's __getattr__ and __setattr__ somewhat. I still believe it would be better to be explicit about dynamic properties, and will try to find some time to make that happen.
Changed a couple exec() calls to setattr() calls instead, which may be safer 
if not faster, but definitely more explicit and easier to understand.

Added back a needed test for deadObject that I removed from the onUpdate
callback, and changed the test in update() back to testing for the deadObject
instead of putting it in a try block.