Added makeDynamicProperty() function to dabo.ui, which creates a dynamic property definition for the passed normal property object. I wanted to put this in propertyHelperMixin but it doesn't appear to me that dynamic properties are really at all useful outside of a UI context (indeed, everything else having to do with dynamic props is defined inside dPemMixin...).
Explicitly defined Dynamic properties for most of the existing properties
defined in dPemMixin. Removed the automagical __getattr__/__setattr__ 
overrides.

This dynamism is really cool, and I'm sorry that we aren't seeing completely
eye-to-eye on it. But here are some things that "just work" with my
implementation:

1) Run SimpleFormWithBizobj, and in the command window type:
> def dfb(biz):
>    return biz.ccity.strip() == "Arecibo"
    
> self.PageFrame.Pages[2].setAll("DynamicFontBold", (dfb, self.getBizobj()))

Now, go to the edit page and scroll through. The setAll() call just worked without
any changes needed. Something is keeping it working with the original implementation
but I admit that I don't see why, at first glance.

2) Notice when you are typing in the command window, that the dynamic properties
show up in the list. This is really what I mean when I talk about things being
documented.

3) Passing dynamic props to the constructor is the same as defining them in the class
txt = dabo.ui.dTextBox(self, DynamicEnabled=myFunc)
txt.DynamicFontItalic = myFunc

Both of the above calls work as usual WRT the Dabo Property Interface.

Granted, there's more work involved when defining a new class (gotta remember to tell
Python about the dynamic props). But after all is said, done, and considered, it is 
my genuine hope and belief that this was the right way to go. But I am still sorry for 
stepping on toes.

I'll follow through with making other properties (those outside dPemMixin) dynamic as
well. 

There are things to consider in the future, such as:

+ Do we make a DynamicValue?
+ If so, does it override the value gotten from FieldSource, if set?
+ What bearing does this have on the dynamism already present in, e.g.,
the setting of dynamic values from non-bizobj data bindings?

Anyway, thanks for putting in the Dynamic stuff, Ed. Once we take care of the refresh
overhead - perhaps by using the timer/observer-based approach I talked about this 
morning - this is really going to be killer.