Table creation now works for MySQL also.
The class dAutoBizobj is a bizobj that has a dTable.

The function AutoCreateTables() will try to create all the tables that have been setup with dAutoBizobj objects. If it can't create the tables, a message pops up for the user and the db admin. The db admin can enter the username and password for access to create tables or the db admin can manualy execute the queries in the queries.sql.

Also, having autotables for more than database or even more than one host does not messup or crash, it is handled.