Converted the dTimer from a bitmap to a non-visible control based directly on wx.Timer. This makes it fit in much better with dabo.ui.callEvery(), which is much more convenient to use in many cases.
Fixed a problem in eventMixin whereby an object without a GetParent() method could cause an error.

Began coding for the Hover property, which will allow for easier visual coding of MouseOver events. It isn't fully implemented yet, so don't try working with it.

Fixed a bug in dPanel whereby Resize events were not propagated to subclasses when Buffered = True.