Added defaults to dGrid.getValue(). If no row or col is specified, it defaults to the current row/column, respectively.
Added fontMetricFromDrawObject() to dabo.ui. This enables simple calculation of the width/height of a drawn text object.

Cleaned up the property methods and definitions of dTextBox to make them alphabetized, and to make the strings localized. Also relaxed the values accepted for the Alignment property.

Added a fitToSizer() method to dPemMixin. This will resize the object to fit the size required by the sizer represented in its Sizer property. You can also pass in extraWidth and extraHeight values to adjust the size further.

Added a call to flushValue() to the _setValue() method for data controls in the event that the value didn't change. I was running into cases where data binding properties were set up after the initial value was set, and there was no way to propagate the data binding without doing something silly like changing the value to something random and then changing it back.

Added 'Alignment' to the pre-init properties of dCheckBox. Cleaned up the property methods and definitions to make them alphabetized, and to make the strings localized. Also relaxed the values accepted for the Alignment property.

Added an option to dEditor.changeFontSize() to accept strings in the format "+n" or "-n". These will increment or decrement the font size by 'n' points.