Refactored the call to the command window from dApp to uiApp.
Fixed a bug in dGrid that would throw errors when a grid was either being created or destroyed, and the calls to paint the headers were out of sync with the current column count.

Fixed a bug in dSizerMixin that would always report the BorderSides property of an object as ['All'], even when it was set to only be specific sides.