Added setAfter() and setAfterInterval() to dabo.ui. These work the same as callAfter() and callAfterInterval(), except you use them to set properties to particular values.
Added the following properties to dGrid: MovableColumns, ResizableColumns, and ResizableRows. When any of these are set to False, the user cannot move/resize the column or row.

Re-arranged the property definitions to be alphabetical. 

Refactored all of the Dynamic* property calls to their own area after the property declarations. IMO, this makes the property definition section *much* easier to read, as it's easier to see each property defintion by itself, and find the one you want more quickly. If this is acceptable to you, Paul, I propose we adopt this as the standard in our property definition sections of code.