Re-added the code to have the grid show None values using grid.NoneDisplay.
Refactored NoneDisplay out of uiApp and into dApp, since there isn't any
ui-specific code needed here. 

Made dGrid's NoneDisplay settable, and if set it overrides the dApp value,
otherwise it pulls dApp's value.

Oh, and I changed the default NoneDisplay to "< None >", because that is
my preference, but this doesn't really matter that much so if it really
bugs, feel free to change it to something else.