Continued the project of getting font behaviors and definitions encapsulated in dFont.
Set up an event so that the dFont instance can notify any 
controls that are using this font that the properties have changed.
This is dEvents.FontPropertiesChanged.

Modified the various font property getters/setters in dPemMixin,
dTreeView, and dGrid to better define and use the dFont instances. 
Previously, a new dFont instance was getting created simply by 
querying any of the font properties. Now, we create and reuse the 
dFont instance and use it as a full proxy to the wxFont object. 

Removed the NativeObject property, as it isn't necessary for user
code to access this directly. They can still get to it using the
_nativeFont attribute, however.

I'll commit this change and see if the dynamic prop demo works 
correctly on Mac now. If not, I've got a better base from which to
fix the problem now.