Fixed  a bug in dabo.ui.saveScreenShot() that only happened if you passed both the path and imgType.
Hacked a bit to get getCaptureBitmap() to work well on GTTK, but it's still
a bit of a mess. GTK doesn't capture the title bar, menu bar or frame
decorations of a window, and I tried to compensate the sizing accordingly.
Individual controls don't seem to work at all. I don't have more time to
play around with this now, but this is hardly a show-stopper for release.