Changed the Dabo trunk version to '0.7a', where the 'a' is shorthand for 'alpha'. All day-to-day development will happen on the 0.7a trunk.
Here are the relevant subversion URL's:

Development (trunk) repo:
svn://paulmcnett.com/dabo/trunk

Stable repo:
svn://paulmcnett.com/dabo/branches/stable

We are attempting to maintain two branches with each release: the 
development branch which will become the next stable version, and the
stable branch will is the prior development version. 

We'll only backport important fixes to the stable branch, not new and
improved functionality. Please everyone help keep an eye on this, and
let us know if we missed backporting an important bugfix.

We still encourage everyone who is using Dabo to stay current with the
daily updates, but getting this mechanism in place now will help us 
once we pass the 1.0 milestone.