Added tests for doDefault() and super() to dObject, so we can test them side by side.
Altered autosuper to not raise an AttributeError if the superclass doesn't
have the method defined, to match the behavior of doDefault(). IMO, this
is better because then people can call self.super() without having to 
remember if they defined the method or if there is a parent method to be
called. If there is no super method, the return value will be None.