Previously, the default directory name and file name for the user settings file was derived from the same thing: app.getAppInfo("appShortName"). Removed that from the file name portion. So now, by default, there will be one settings file (user.ini) inside <home>/.<appShortName>. As always, appShortName defaults to "DaboApplication", so by default for bare Dabo applications, the settings file will be in <home>/.daboapplication/user.ini.
For AppWizard-generated applications, the settings file will vary depending
on the name of the database. So, if you generated your app based on a 
database called 'zipcodes', the user settings will be in <home>/.zipcodes/user.ini. You can change this simply, in your generated App.py file, by adding
code such as the following to initProperties():
	self.UserSettingProvider.SettingsDirectoryName = "Dabo"
	self.UserSettingProver.SettingsFileName = "ZipCodes.ini"