Added a test to xmltodict, which starts with a xmldict, calls dicttoxml, and then calls xmltodict on that, to test the double-backslash thing.
The test fails, because we don't end up with a dict that matches the original
dict, because of the replacing of "\\" with "\\\\". If you comment line 168
and uncomment line 169, the test succeeds.

This leads me to believe that xmltodict is correct without the conversion (and
the fix for windows pathing should happen elsewhere), but I'm not sure yet
so for now at least we can play around with the test. No matter what we end
up with, ending up with an identical dict is a requirement!