Created the dPref class. This is an attempt to make the handling of preferences much simpler and more direct. It is based on the concept of nested preferences, so that by running code like:
pref = dabo.dPref(level="MyCoolApp")
pref.positions.MainWindow.Top = 33

will create several nested objects: pref, pref.positions, and pref.positions.MainWindow. The value of 33 is automaticall persisted to a SQLite database in the home dabo directory (typically ~/.dabo/DaboPreferences.db on POSIX systems). No explicit commit is needed, unless the AutoPersist property is set to False. In that case, in order to persist preference settings to the database, you need to call the preference object's persist() method explicitly.

This is still very much a work-in-progress, and needs a lot of refinement before it is ready to replace the current userSettings system.