Added two new properties to dObject: BasePrefKey and PreferenceManager. 'BasePrefKey' defines the identifying key used to store preferences for that particular object; 'PreferenceManager' is a reference to a dPref object whose key is automatically set to self.BasePrefKey.
Cleaned up some poor naming choices for atts in dPref, and added a restriction that if there is no key value set, the values set in the object will not be persisted to the database. There is simply too much potential for harm when multiple applications are writing preferences to the same namespace. It is much better to require a key that can be as complex as you like; the more complex, the less chance of a conflict. For example, the key for the Class Designer app might be something like 'dabo.ide.ClassDesigner.app', which is unlikely to clash with any other app's preferences.