Added basic type coercion to the _setValue() method of dDataControlMixin. This will help eliminate spurious mis-matched type errors.
Also found a case where setting a prop in a data-bound control before it ever receives focus can cause the setting to be cleared. This was the case in the demo/simpleDemo3.py file, where the setting of the slider's Value to 42 was wiped out and re-set to zero, due to the underlying bound label.Caption property being empty. This is now fixed.

Added one more possible way of determining mouse position in the getEventData() method of uiwx/__init__.py.