Re-vamped the dToggleButton class to use the wx.lib.buttons.GenBitmapTextToggleButton instead, allowing for both text and images on the button.
Refactored out the standard image-setting properties into dImageMixin, and that is now used by the different controls that display images. dBitmapButton is still a bit problematic, as it has several images it uses, so we might be able to refactor this a bit better.

Fixed some localization strings in dForm, and added a test for the borderless form class.