Refactored the old DataSet class out of dCursorMixin and into its own Dabo base class, named 'dDataSet'.
Added a 'filter()' method that will accept some basic info, such as field and value, and return a filtered data set. This is working, but I plan on adding a function to clear the filter and return the original data set, as well as add a filter to a filter, etc.