Removed the 'txt' attribute from the dNode class. The Caption is now determined by querying the tree directly, which makes for better syncing when updating the Caption. Added 'IsRootNode' property that reflects just what the name says.
Added the 'ShowLines' property that controls the display of lines connecting the nodes. Changed the class to default to True for ShowLines, ShowButtons, ShowRootNode, and ShowRootNodeLines.

Fixed a bug in the expandAll() and collapseAll() methods when the root node was hidden.

Improved the self-test code to better illustrate the various properties.