Added the drawGradient() method. This will draw a gradient that varies between the two colors specified in either the horizontal or vertical direction, depending on the orientation of the gradient.
There are still some problems with getting the drawing to stay 'behind' the other elements, which are sometimes drawn before the gradient, but perhaps some enterprising soul will figure out how to deal with this. The code to draw the gradients is taken from Andrea Gavana's FoldPanelBar class, where he uses it to draw the gradients on the caption bars.