Added dabo.ui.getFormMousePosition() method. This returns the mouse position relative to the active form.
Began working on a problem with wxPython's FindWindowAtPoint() method, which seems to report the wrong window when a frame is positioned to the left and above the frame containing the mouse. If you use the dabo.ui.getMouseObject() method, there will be some extra debugging output.

Added before- and after- hooks to setProperties() in dPemMixin. This was needed in order to remove properties needed by the Class Designer but that don't appear in the runtime version of the classes.

Added a relativeCoordinates(pos) method that returns the position of an absolute screen location relative to the object.

Fixed a problem with forms not handling the CxnName property correctly. Issue Tracker #0201. Added the 'CxnFile' property, which together with CxnName, identifies and opens the connection when the form is run.

Cleaned up all the Dynamic* declarations in dFormMixin.